updates for 13 TeV

+++ EventAnalyzer/src/EventVariableOSSF.cpp
@@ -23,12 +23,12 @@ bool EventVariableOSSF::calculate(BaseHandler* handler)
     int nPlus = 0;
     int nMinus = 0;
     for(int j = 0; j < (int)v.size(); j++){
-      double charge1 = 0;
+      int charge1 = 0;
       bool charge1set = (v[j])->getVariable("CHARGE",charge1);
       if(!charge1set)continue;
       charge1 > 0 ? nPlus++ : nMinus++;
       for(int k = j+1; k < (int)v.size(); k++){
-	double charge2 = 0;
+	int charge2 = 0;
 	bool charge2set = (v[k])->getVariable("CHARGE",charge2);
 	if(!charge2set)continue;
 	if(charge1*charge2 < 0){
